Changed Biodata edit form

+++ Data/CMdm.Data/DAC/DqQueDAC.cs
@@ -360,6 +360,11 @@ namespace CMdm.Data.DAC
         {
             using (var db = new AppDbContext())
             {
+                //The property 'AUTHORISED' is part of the object's key information and cannot be modified.
+                //entry2.AUTHORISED = "A";
+                //db.CDMA_INDIVIDUAL_NEXT_OF_KIN.Attach(entry2);
+                //db.Entry(entry2).State = System.Data.Entity.EntityState.Modified;
+                int noOfRowUpdated = 0;
                 foreach (var item in modifiedrecords)
                 {
                     var queitem = db.MdmDqRunExceptions.FirstOrDefault(a => a.EXCEPTION_ID == item.EXCEPTION_ID);
@@ -377,15 +382,15 @@ namespace CMdm.Data.DAC
                             var oldrecord = db.CDMA_INDIVIDUAL_BIO_DATA.FirstOrDefault(a => a.CUSTOMER_NO == item.CUST_ID && a.AUTHORISED == "A");
                             if (entry != null)
                             {
-                                entry.AUTHORISED = "A";
-                                db.CDMA_INDIVIDUAL_BIO_DATA.Attach(entry);
-                                db.Entry(entry).State = System.Data.Entity.EntityState.Modified;
                                 if (oldrecord != null)
                                 {
                                     db.CDMA_INDIVIDUAL_BIO_DATA.Remove(oldrecord);
                                 }
-                                
+
                                 db.SaveChanges();
+                                noOfRowUpdated = db.Database.ExecuteSqlCommand("Update CDMA_INDIVIDUAL_BIO_DATA set AUTHORISED='A', AUTHORISED_BY = " + userId + ", AUTHORISED_DATE=sysdate  where customer_no = '" + item.CUST_ID + "' and AUTHORISED ='U' ");
+
+
                             }
                             break;
                         case "CDMA_CUSTOMER_INCOME":
@@ -393,16 +398,14 @@ namespace CMdm.Data.DAC
                             var oldrecord1 = db.CDMA_CUSTOMER_INCOME.FirstOrDefault(a => a.CUSTOMER_NO == item.CUST_ID && a.AUTHORISED == "A");
                             if (entry1 != null)
                             {
-                                entry1.AUTHORISED = "A";
-                                db.CDMA_CUSTOMER_INCOME.Attach(entry1);
-                                db.Entry(entry1).State = System.Data.Entity.EntityState.Modified;
-
                                 if (oldrecord1 != null)
                                 {
                                     db.CDMA_CUSTOMER_INCOME.Remove(oldrecord1);
                                 }
-                                
+
                                 db.SaveChanges();
+                                noOfRowUpdated = db.Database.ExecuteSqlCommand("Update CDMA_CUSTOMER_INCOME set AUTHORISED='A', AUTHORISED_BY = " + userId + ", AUTHORISED_DATE=sysdate  where customer_no = '" + item.CUST_ID + "' and AUTHORISED ='U' ");
+
                             }
                             break;
                         case "CDMA_INDIVIDUAL_NEXT_OF_KIN":
@@ -415,15 +418,8 @@ namespace CMdm.Data.DAC
                                     db.CDMA_INDIVIDUAL_NEXT_OF_KIN.Remove(oldrecord2);
                                 }
 
-                                    //The property 'AUTHORISED' is part of the object's key information and cannot be modified.
-                                //entry2.AUTHORISED = "A";
-                                //db.CDMA_INDIVIDUAL_NEXT_OF_KIN.Attach(entry2);
-                                //db.Entry(entry2).State = System.Data.Entity.EntityState.Modified;
-
-
-
                                 db.SaveChanges();
-                                int noOfRowUpdated = db.Database.ExecuteSqlCommand("Update CDMA_INDIVIDUAL_NEXT_OF_KIN set AUTHORISED='A', AUTHORISED_BY = " + userId + ", AUTHORISED_DATE=sysdate  where customer_no = '" + item.CUST_ID + "' and AUTHORISED ='U' ");
+                                noOfRowUpdated = db.Database.ExecuteSqlCommand("Update CDMA_INDIVIDUAL_NEXT_OF_KIN set AUTHORISED='A', AUTHORISED_BY = " + userId + ", AUTHORISED_DATE=sysdate  where customer_no = '" + item.CUST_ID + "' and AUTHORISED ='U' ");
 
                             }
                             break;
@@ -432,15 +428,14 @@ namespace CMdm.Data.DAC
                             var oldrecord3 = db.CDMA_FOREIGN_DETAILS.FirstOrDefault(a => a.CUSTOMER_NO == item.CUST_ID && a.AUTHORISED == "A");
                             if (entry3 != null)
                             {
-                                entry3.AUTHORISED = "A";
-                                db.CDMA_FOREIGN_DETAILS.Attach(entry3);
-                                db.Entry(entry3).State = System.Data.Entity.EntityState.Modified;
-                                if(oldrecord3 != null)
+                                if (oldrecord3 != null)
                                 {
                                     db.CDMA_FOREIGN_DETAILS.Remove(oldrecord3);
                                 }
-                                
+
                                 db.SaveChanges();
+                                noOfRowUpdated = db.Database.ExecuteSqlCommand("Update CDMA_FOREIGN_DETAILS set AUTHORISED='A', AUTHORISED_BY = " + userId + ", AUTHORISED_DATE=sysdate  where customer_no = '" + item.CUST_ID + "' and AUTHORISED ='U' ");
+
                             }
                             break;
                         case "CDMA_JURAT":
@@ -448,15 +443,14 @@ namespace CMdm.Data.DAC
                             var oldrecord4 = db.CDMA_JURAT.FirstOrDefault(a => a.CUSTOMER_NO == item.CUST_ID && a.AUTHORISED == "A");
                             if (entry4 != null)
                             {
-                                entry4.AUTHORISED = "A";
-                                db.CDMA_JURAT.Attach(entry4);
-                                db.Entry(entry4).State = System.Data.Entity.EntityState.Modified;
                                 if (oldrecord4 != null)
                                 {
                                     db.CDMA_JURAT.Remove(oldrecord4);
                                 }
-                                
+
                                 db.SaveChanges();
+                                noOfRowUpdated = db.Database.ExecuteSqlCommand("Update CDMA_JURAT set AUTHORISED='A', AUTHORISED_BY = " + userId + ", AUTHORISED_DATE=sysdate  where customer_no = '" + item.CUST_ID + "' and AUTHORISED ='U' ");
+
                             }
                             break;
                         case "CDMA_EMPLOYMENT_DETAILS":
@@ -464,15 +458,14 @@ namespace CMdm.Data.DAC
                             var oldrecord5 = db.CDMA_EMPLOYMENT_DETAILS.FirstOrDefault(a => a.CUSTOMER_NO == item.CUST_ID && a.AUTHORISED == "A");
                             if (entry5 != null)
                             {
-                                entry5.AUTHORISED = "A";
-                                db.CDMA_EMPLOYMENT_DETAILS.Attach(entry5);
-                                db.Entry(entry5).State = System.Data.Entity.EntityState.Modified;
-                                if (oldrecord5 != null)
+                                if(oldrecord5 != null)
                                 {
                                     db.CDMA_EMPLOYMENT_DETAILS.Remove(oldrecord5);
                                 }
-                                
+
                                 db.SaveChanges();
+                                noOfRowUpdated = db.Database.ExecuteSqlCommand("Update CDMA_EMPLOYMENT_DETAILS set AUTHORISED='A', AUTHORISED_BY = " + userId + ", AUTHORISED_DATE=sysdate  where customer_no = '" + item.CUST_ID + "' and AUTHORISED ='U' ");
+
                             }
                             break;
                         case "CDMA_TRUSTS_CLIENT_ACCOUNTS":
@@ -480,15 +473,14 @@ namespace CMdm.Data.DAC
                             var oldrecord6 = db.CDMA_TRUSTS_CLIENT_ACCOUNTS.FirstOrDefault(a => a.CUSTOMER_NO == item.CUST_ID && a.AUTHORISED == "A");
                             if (entry6 != null)
                             {
-                                entry6.AUTHORISED = "A";
-                                db.CDMA_TRUSTS_CLIENT_ACCOUNTS.Attach(entry6);
-                                db.Entry(entry6).State = System.Data.Entity.EntityState.Modified;
                                 if (oldrecord6 != null)
                                 {
                                     db.CDMA_TRUSTS_CLIENT_ACCOUNTS.Remove(oldrecord6);
                                 }
-                                
+
                                 db.SaveChanges();
+                                noOfRowUpdated = db.Database.ExecuteSqlCommand("Update CDMA_TRUSTS_CLIENT_ACCOUNTS set AUTHORISED='A', AUTHORISED_BY = " + userId + ", AUTHORISED_DATE=sysdate  where customer_no = '" + item.CUST_ID + "' and AUTHORISED ='U' ");
+
                             }
                             break;
                         case "CDMA_AUTH_FINANCE_INCLUSION":
@@ -496,11 +488,14 @@ namespace CMdm.Data.DAC
                             var oldrecord7 = db.CDMA_AUTH_FINANCE_INCLUSION.FirstOrDefault(a => a.CUSTOMER_NO == item.CUST_ID && a.AUTHORISED == "A");
                             if (entry7 != null)
                             {
-                                entry7.AUTHORISED = "A";
-                                db.CDMA_AUTH_FINANCE_INCLUSION.Attach(entry7);
-                                db.Entry(entry7).State = System.Data.Entity.EntityState.Modified;
-                                db.CDMA_AUTH_FINANCE_INCLUSION.Remove(oldrecord7);
+                                if (oldrecord7 != null)
+                                {
+                                    db.CDMA_AUTH_FINANCE_INCLUSION.Remove(oldrecord7);
+                                }
+
                                 db.SaveChanges();
+                                noOfRowUpdated = db.Database.ExecuteSqlCommand("Update CDMA_AUTH_FINANCE_INCLUSION set AUTHORISED='A', AUTHORISED_BY = " + userId + ", AUTHORISED_DATE=sysdate  where customer_no = '" + item.CUST_ID + "' and AUTHORISED ='U' ");
+
                             }
                             break;
                         case "CDMA_ADDITIONAL_INFORMATION":
@@ -508,15 +503,14 @@ namespace CMdm.Data.DAC
                             var oldrecord8 = db.CDMA_ADDITIONAL_INFORMATION.FirstOrDefault(a => a.CUSTOMER_NO == item.CUST_ID && a.AUTHORISED == "A");
                             if (entry8 != null)
                             {
-                                entry8.AUTHORISED = "A";
-                                db.CDMA_ADDITIONAL_INFORMATION.Attach(entry8);
-                                db.Entry(entry8).State = System.Data.Entity.EntityState.Modified;
                                 if (oldrecord8 != null)
                                 {
                                     db.CDMA_ADDITIONAL_INFORMATION.Remove(oldrecord8);
                                 }
-                                
+
                                 db.SaveChanges();
+                                noOfRowUpdated = db.Database.ExecuteSqlCommand("Update CDMA_ADDITIONAL_INFORMATION set AUTHORISED='A', AUTHORISED_BY = " + userId + ", AUTHORISED_DATE=sysdate  where customer_no = '" + item.CUST_ID + "' and AUTHORISED ='U' ");
+
                             }
                             break;
                         default:
